Mainにbottom navigation を移行
