+++ app/src/main/AndroidManifest.xml
@@ -14,13 +14,7 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme"
         tools:ignore="AllowBackup,GoogleAppIndexingWarning">
-        <activity android:name=".BottomNavigation">
-        <intent-filter>
-            <action android:name="android.intent.action.MAIN" />
-
-            <category android:name="android.intent.category.LAUNCHER" />
-        </intent-filter>
-        </activity>
+        <activity android:name=".BottomNavigation"/>
         <activity
             android:name=".MainPostAddActivity"
             android:theme="@style/AppTheme.NoActionBar"/>
@@ -33,6 +27,11 @@
         <activity
             android:name=".SignInActivity"
             android:theme="@style/AppTheme.NoActionBar">
+            <intent-filter>
+                <action android:name="android.intent.action.MAIN" />
+
+                <category android:name="android.intent.category.LAUNCHER" />
+            </intent-filter>
         </activity>
         <activity android:name=".StartUpActivity" />
