+++ app/src/main/java/jp/ac/asojuku/st/chirusapo/MainActivity.kt
@@ -1,20 +1,23 @@
 package jp.ac.asojuku.st.chirusapo
 
-import android.content.Intent
+import android.net.Uri
 import android.os.Bundle
-import com.google.android.material.floatingactionbutton.FloatingActionButton
 import androidx.navigation.findNavController
-import androidx.navigation.ui.AppBarConfiguration
-import androidx.navigation.ui.navigateUp
-import androidx.navigation.ui.setupActionBarWithNavController
-import androidx.navigation.ui.setupWithNavController
-import androidx.drawerlayout.widget.DrawerLayout
-import com.google.android.material.navigation.NavigationView
 import androidx.appcompat.app.AppCompatActivity
 import androidx.appcompat.widget.Toolbar
 import android.view.Menu
+import androidx.navigation.ui.*
+import kotlinx.android.synthetic.main.content_main.*
 
-class MainActivity : AppCompatActivity() {
+class MainActivity : AppCompatActivity(),
+    HomeFragment.OnFragmentInteractionListener,
+    ChildFragment.OnFragmentInteractionListener,
+    CalendarFragment.OnFragmentInteractionListener,
+    AlubamFragment.OnFragmentInteractionListener,
+    DressFragment.OnFragmentInteractionListener {
+    override fun onFragmentInteraction(uri: Uri) {
+
+    }
 
     private lateinit var appBarConfiguration: AppBarConfiguration
 
@@ -24,24 +27,29 @@ class MainActivity : AppCompatActivity() {
         val toolbar: Toolbar = findViewById(R.id.toolbar)
         setSupportActionBar(toolbar)
 
-        val fab: FloatingActionButton = findViewById(R.id.fab)
-        fab.setOnClickListener {
-            val intent = Intent(this, MainPostAddActivity::class.java)
-            startActivity(intent)
-        }
-        val drawerLayout: DrawerLayout = findViewById(R.id.drawer_layout)
-        val navView: NavigationView = findViewById(R.id.nav_view)
-        val navController = findNavController(R.id.nav_host_fragment)
+//        val fab: FloatingActionButton = findViewById(R.id.fab)
+//        fab.setOnClickListener {
+//            val intent = Intent(this, MainPostAddActivity::class.java)
+//            startActivity(intent)
+//        }
+//        val drawerLayout: DrawerLayout = findViewById(R.id.drawer_layout)
+//        val navView: NavigationView = findViewById(R.id.nav_view)
+//        val navController = findNavController(R.id.nav_host_fragment)
         // Passing each menu ID as a set of Ids because each
         // menu should be considered as top level destinations.
-        appBarConfiguration = AppBarConfiguration(
-            setOf(
-//                R.id.nav_home, R.id.nav_group_participation, R.id.nav_create_group,
-                R.id.nav_config, R.id.nav_logout, R.id.nav_create_group
-            ), drawerLayout
-        )
-        setupActionBarWithNavController(navController, appBarConfiguration)
-        navView.setupWithNavController(navController)
+//        appBarConfiguration = AppBarConfiguration(
+//            setOf(
+////                R.id.nav_home, R.id.nav_group_participation, R.id.nav_create_group,
+//                R.id.nav_config, R.id.nav_logout, R.id.nav_create_group
+//            ), drawerLayout
+//        )
+//        setupActionBarWithNavController(navController, appBarConfiguration)
+//        navView.setupWithNavController(navController)
+
+//        setContentView(R.layout.activity_bottom_navigation)
+
+        val navBottomController = findNavController(R.id.nav_host_fragment)
+        NavigationUI.setupWithNavController(navigation, navBottomController)
     }
 
     override fun onCreateOptionsMenu(menu: Menu): Boolean {
@@ -50,8 +58,8 @@ class MainActivity : AppCompatActivity() {
         return true
     }
 
-    override fun onSupportNavigateUp(): Boolean {
-        val navController = findNavController(R.id.nav_host_fragment)
-        return navController.navigateUp(appBarConfiguration) || super.onSupportNavigateUp()
-    }
+//    override fun onSupportNavigateUp(): Boolean {
+//        val navController = findNavController(R.id.nav_host_fragment)
+//        return navController.navigateUp(appBarConfiguration) || super.onSupportNavigateUp()
+//    }
 }
